implemeted authorization using gates

+++ app/Http/Controllers/QuestionsController.php
@@ -68,7 +68,11 @@ class QuestionsController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function edit(Question $question)
-    {
+    {   
+        if (\Gate::denies('update-question', $question)){
+            abort(403, "Access denied");
+        }
+        
         return view('questions.edit', compact('question'));
     }
 
@@ -80,7 +84,11 @@ class QuestionsController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function update(AskQuestionRequest $request, Question $question)
-    {
+    {   
+
+        if (\Gate::denies('update-question', $question)){
+            abort(403, "Access denied");
+        }
         $question->update($request->only('title', 'body'));
 
         return redirect()->route('questions.index')->with('success', "Your Question is updated!");
@@ -93,7 +101,10 @@ class QuestionsController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function destroy(Question $question)
-    {
+    {   
+        if (\Gate::denies('delete-question', $question)){
+            abort(403, "Access denied");
+        }
         $question->delete();
 
         return redirect()->route('questions.index')->with('success', " The question is deleted!");
